Fix deserialisation issues
On time fields, I used camel case when writing instead of snake case,
which is what I used when reading.

I also wrote explicit nulls to the JSON data, but forgot to treat nulls
as absent when I read them, causing type issues. Add a local utility
function to treat nulls the same as absent values, and use that instead
of a direct .get on the map of fields