Targeting logic and ClusterHandling
Add TargetResolution logic which identifies targets based on config
Add ClusterHandling trait which builds a graph which connects personal
and shared queues, directs notifications to the appropriate route based
on target, and acts as a Source for the existing notification graph

This involved building a partial graph which acts as a Source but is
relatively complex in itself.

NotificationReading/Reader is now NotificationSourceMerging/Merger, to
indicate that its actual function is essentially to take a load of other
sources, possibly from config, and merge them into a single source (as
well as providing the akka Source logic for merging them and supervising
the resulting Source)

In Main we now create the ClusterHandler directly using config for
individual and cluster queues, as well as personal client ID and peer
IDs, and then collect that and other sources together by continuing to
accept configured classes but dropping QueueHandler from the configured
classes as it's now part of the ClusterHandler.

Configured classes are now empty by default but this allows more
extensibility.