Make cluster handler a separate component
It can't simply be a notification source as actually it's the component
which should provide the actor ref input.

Instead of treating it as just another notification source, acknowledge
it as special and factor it into the NotificationStreaming graph;
StreamingApp now takes a cluster handler, which continues to come
equipped with its own notification source, and a separate source for
personal items which shouldn't go through the cluster system.

For the moment, the cluster handler is set up in Main with all
configured sources, including the cluster queue of course, and the
"direct" source is only the personal queue.