Move RestartSource/Sink directly onto AMQP comps
Where RestartSource was, a failure of the AMQP source would cause failure
to be triggered to the merged source, but that would not result in
failure of the merged source, as it'd wait for all upstream elements to complete.

Also make this eager to fail, meaning any upstream failure will fail the
merged source - however since the RestartSource now applies to AMQP, it
shouldn't fail unless there's a prolonged error.

Similarly, RestartSink has been moved into the AMQPWriter, for the same
reasons.